+++ src/commands/block/transactions.ts
@@ -65,7 +65,7 @@ export default class extends ProfileCommand {
 
         this.spinner.start();
         const profile = this.getProfile(options);
-        const height = new HeightResolver().resolve(options);
+        const height = new HeightResolver().resolve(options).toString();
 
         let pageSize = options.pageSize || 10;
         if (pageSize < 10) {