+++ src/commands/transaction/namespace.ts
@@ -16,17 +16,12 @@
  *
  */
 import {command, metadata, option} from 'clime';
-import {Deadline, NamespaceRegistrationTransaction} from 'nem2-sdk';
-import * as readlineSync from 'readline-sync';
-import {OptionsResolver} from '../../options-resolver';
+import {Deadline, NamespaceRegistrationTransaction, NamespaceRegistrationType} from 'nem2-sdk';
 import {AnnounceResolver} from '../../resolvers/announce.resolver';
 import {DurationResolver} from '../../resolvers/duration.resolver';
 import {MaxFeeResolver} from '../../resolvers/maxFee.resolver';
-import {
-    AnnounceTransactionFieldsTable,
-    AnnounceTransactionsCommand,
-    AnnounceTransactionsOptions,
-} from '../announce.transactions.command';
+import {NamespaceNameResolver, NamespaceTypeResolver} from '../../resolvers/namespace.resolver';
+import {AnnounceTransactionFieldsTable, AnnounceTransactionsCommand, AnnounceTransactionsOptions} from '../announce.transactions.command';
 
 export class CommandOptions extends AnnounceTransactionsOptions {
     @option({
@@ -77,33 +72,22 @@ export default class extends AnnounceTransactionsCommand {
         const profile = this.getProfile(options);
         const account = profile.decrypt(options);
 
-        options.name = OptionsResolver(options,
-            'name',
-            () => undefined,
-            'Enter namespace name: ');
-
-        if (!options.rootnamespace && readlineSync.keyInYN('Do you want to create a root namespace?')) {
-            options.rootnamespace = true;
-        }
-        let duration;
-        if (!options.rootnamespace) {
-            options.subnamespace = true;
-            options.parentName = OptionsResolver(options,
-                'parentName',
-                () => undefined,
-                'Enter the parent namespace name: ');
-        } else {
-            duration = new DurationResolver().resolve(options);
-        }
+        const namespaceId = new NamespaceNameResolver()
+            .resolve(options, undefined, 'Enter the namespace name: ', 'name');
+        const name = namespaceId.fullName ? namespaceId.fullName : '';
+        const namespaceType = new NamespaceTypeResolver().resolve(options);
         const maxFee = new MaxFeeResolver().resolve(options);
 
         let transaction: NamespaceRegistrationTransaction;
-        if (options.rootnamespace) {
+        if (namespaceType === NamespaceRegistrationType.RootNamespace) {
+            const duration = new DurationResolver().resolve(options);
             transaction = NamespaceRegistrationTransaction.createRootNamespace(
-                Deadline.create(), options.name, duration, profile.networkType, maxFee);
+                Deadline.create(), name, duration, profile.networkType, maxFee);
         } else {
+            const parentName = new NamespaceNameResolver()
+                .resolve(options, undefined, 'Enter the parent namespace name: ', 'parentName');
             transaction = NamespaceRegistrationTransaction.createSubNamespace(
-                Deadline.create(), options.name, options.parentName, profile.networkType, maxFee);
+                Deadline.create(), name, parentName, profile.networkType, maxFee);
         }
         const signedTransaction = account.sign(transaction, profile.networkGenerationHash);