+++ src/resolvers/amount.resolver.ts
@@ -17,7 +17,7 @@ export class AmountResolver implements Resolver {
      * @param {string} altText - Alternative text.
      * @returns {UInt64}
      */
-    resolve(options: ProfileOptions, secondSource?: Profile, altText?: string): any {
+    resolve(options: ProfileOptions, secondSource?: Profile, altText?: string): UInt64 {
         const resolution = OptionsResolver(options,
         'amount',
         () =>  undefined,