+++ src/resolvers/default.resolver.ts
@@ -9,13 +9,13 @@ import {Resolver} from './resolver';
 export class DefaultResolver implements Resolver {
 
     /**
-     * Resolves an action provided by the user.
+     * Resolves if an account has to be set as default.
      * @param {CreateProfileOptions} options - Command options.
      * @param {Profile} secondSource - Secondary data source.
      * @param {string} altText - Alternative text.
-     * @returns {number}
+     * @returns {boolean}
      */
-    resolve(options: CreateProfileOptions, secondSource?: Profile, altText?: string): any {
+    resolve(options: CreateProfileOptions, secondSource?: Profile, altText?: string): boolean {
         if (!options.default && readlineSync.keyInYN('Do you want to set the account as the default profile?')) {
             options.default = true;
         }