+++ src/resolvers/payload.resolver.ts
@@ -1,4 +1,4 @@
-import {TransactionMapping} from 'nem2-sdk';
+import {Transaction, TransactionMapping} from 'nem2-sdk';
 import {ProfileOptions} from '../commands/profile.command';
 import {Profile} from '../models/profile';
 import {OptionsResolver} from '../options-resolver';
@@ -16,7 +16,7 @@ export class PayloadResolver implements Resolver {
      * @param {string} altText - Alternative text.
      * @returns {Transaction}
      */
-    resolve(options: ProfileOptions, secondSource?: Profile, altText?: string): any {
+    resolve(options: ProfileOptions, secondSource?: Profile, altText?: string): Transaction {
         const resolution = OptionsResolver(options,
             'payload',
             () => undefined,