+++ test/resolvers/message.resolver.spec.ts
@@ -16,7 +16,7 @@
  *
  */
 import {expect} from 'chai';
-import {MessageResolver, RecipientPublicKeyResolver} from '../../src/resolvers/message.resolver';
+import {MessageResolver} from '../../src/resolvers/message.resolver';
 
 describe('Message resolver', () => {
 
@@ -26,22 +26,4 @@ describe('Message resolver', () => {
         expect(new MessageResolver().resolve(profileOptions))
             .to.be.equal('10');
     });
-});
-
-describe('Recipient public key resolver', () => {
-
-    it('should return public account', () => {
-        const recipientPublicKey = '0000000000000000000000000000000000000000000000000000000000000000';
-        const profileOptions = {recipientPublicKey} as any;
-        expect(new RecipientPublicKeyResolver().resolve(profileOptions).publicKey)
-            .to.be.equal(recipientPublicKey);
-    });
-
-    it('should throw error if invalid public key', () => {
-        const recipientPublicKey = '00000';
-        const profileOptions = {recipientPublicKey} as any;
-        expect(() => new RecipientPublicKeyResolver().resolve(profileOptions))
-            .to.throws(Error);
-    });
-
 });